refactor(generation): split baseplateGenerator into focused modules
Decompose the 1435-LOC baseplateGenerator.ts into 9 sibling modules by
computational concern. baseplateGenerator.ts now contains only the public
API (`generateBaseplate`, `buildBaseplateSolid`, `exportBaseplate`,
`BaseplateProbe`) plus the orchestration that ties siblings together.

Extracted modules (all under 400 LOC):

- baseplateBatchOps.ts — `cutInBatches` + `BOOLEAN_BATCH_SIZE`
  Bounded-memory boolean cuts for large grids.
- baseplateCaches.ts — three LRU caches (`pocketTemplate`,
  `meshResult`, `slabWithPockets`) and the public `clearBaseplateCaches`
  / `getBaseplateCacheStats` / `resetBaseplateCacheStats` (re-exported
  from the orchestrator for caller compatibility).
- baseplateCacheKeys.ts — `meshCacheKey`, `slabPocketsCacheKey`.
- baseplatePockets.ts — pocket cutter geometry: `pocketCacheKey`,
  `pocketSection`, `buildPocketCutter`, `buildSimplifiedPocketCutter`,
  `getPocketTemplate`. Also owns `INSET_TOP` / `INSET_MID`.
- baseplateMagnets.ts — `buildMagnetHoles`.
- baseplateConnectors.ts — `buildConnectors` (dovetail tongues + grooves).
- baseplateEdges.ts — `computeBaseplateEdgeLines` + 5 private helpers
  for procedural edge-line generation, plus `EDGE_CORNER_SEGMENTS` and
  `EDGE_CIRCLE_SEGMENTS`.
- baseplateSlab.ts — `buildSlabProfile`, `sanitizeParams`, `tagOp`.
- baseplateSTL.ts — `buildBaseplateSTL` (binary STL emitter with the
  winding-repair JSDoc retained verbatim).

Public surface unchanged:
- generation.worker.ts still imports `clearBaseplateCaches` from
  baseplateGenerator.
- workerContext.ts still imports `getBaseplateCacheStats` /
  `resetBaseplateCacheStats` from baseplateGenerator.
- diagnoseBaseplateWinding.test.ts still imports `buildBaseplateSolid`,
  `clearBaseplateCaches`, `BaseplateProbe` from baseplateGenerator.
- exportHandler.ts still imports `exportBaseplate` from baseplateGenerator.
- generateHandler.ts still imports `generateBaseplate` from
  baseplateGenerator.

baseplateGenerator.ts: 1435 LOC -> 373 LOC. All 10,278 tests pass,
including the 4 baseplate scenario suites (dovetail, split-stress,
winding) that validate the public surface.
